I have added an example dtls-server.rs program that is supposed to implement the dtls-server.c program that will be give a full working demo of the tls library. I hvae specified all the functions being called and the location of the original C files as in the mbed-tls project. I have also tried to classify the methods based on the complexity as it appeared to me.
I propose that we implement the methods in the modules as called in this program so that we may have a running server, rather than haphazardly implementing random files and all the methods defined there. We may then start with the dtls-client program which is calling almost the same methods as the server.